renderer: migrate consumers off CPU renderer; delete factoriax/renderer.py
Phase A items 3.4 + 3.5 + 3.6 bundled (the deletion would break
consumers if shipped without the migration, so they ride together).
This completes the 'single render path' goal of Item 3.

3.4 — headless consumers move from factoriax.renderer.render_pixels
onto factoriax.jax_renderer.JaxRenderer.jit_render_map. Each module
caches its renderer at module scope keyed by tile_px so JIT cache
warm-up costs are paid once per process per size:

- factoriax/observations.py:rgb() — vision-mode obs.
- factoriax/analysis/video.py — trajectory video export.
- factoriax/agentdebugger/main.py + layout.py — replay rendering.
- baselines/basic_skills/train_single.py + skills/train_ppo.py —
  eval video paths.
- factoriax/play/game_ui.py — caught via grep, not in the audit.
  Migrated. Animation in play (water shimmer, machine pulse, belt
  cargo dots) is gone as a v1 trade-off per atlas.layout.md.

3.5 — icon and texture utilities that play / editor / menu /
agentdebugger still need move to factoriax/ui/icons.py: render_item_icon
and its 25+ shape-drawing helpers, PLAYER_COLORS, MACHINE_TO_ITEM,
create_player_texture, create_player_start_icon, create_biter_texture,
create_default_textures, get_textures, build_texture_lookup. New
module is 1344 lines (from the ~1850 lines of the old renderer.py);
the world-rendering code that the JAX renderer now owns is dropped.

3.6 — factoriax/renderer.py is gone. git grep factoriax.renderer
returns zero hits. Docstrings in factoriax_env.py, play/ui.py, and
agentdebugger/main.py that referenced the old module are updated.

Tests: tests/test_renderer.py deleted (purely about the now-removed
procedural renderer); the three load-bearing icon-direction invariants
moved to TestDirectionalIcons in tests/test_machine_icon_coverage.py.
tests/test_inventory.py, tests/play/test_smoke.py, and
tests/agentdebugger/test_replay.py migrated, not deleted.

Pre-existing dtype bug fix: factoriax/jax_renderer.py:_stamp_player
was passing int16,int16,int32 to jax.lax.dynamic_update_slice. The
CPU renderer hid this because nothing ran jit_render_map directly on
a state from build_state/generate_state (they have int16 player
positions). Cast to int32 inside _stamp_player.

Other touches:
- scripts/jax_render_benchmark.py drops its CPU baseline (the function
  it benchmarked is gone). Sample-image save now writes
  jax_map_render.png via JaxRenderer.jit_render_map.
- docs/generate_screenshots.py uses JaxRenderer(tile_px=32).

Phase A is complete: engine schema (achievements + level on ctor),
gymnax-conformant reset, single render path through the JAX renderer,
sprite atlas as the swap point for visual identity. mypy clean
(81 src files), ruff clean, full fast suite green (~1300 tests).

+++ factoriax/jax_renderer.py
@@ -367,10 +367,13 @@ def render_map(
     num_players = state.player_positions.shape[0]
 
     def _stamp_player(i: int, img: jnp.ndarray) -> jnp.ndarray:
-        px = state.player_positions[i, 0]
-        py = state.player_positions[i, 1]
+        # Cast indices to int32 so all three coordinates passed to
+        # ``dynamic_update_slice`` share the same dtype. ``player_positions``
+        # is int16 in level/world states; the literal 0 is int32.
+        px = state.player_positions[i, 0].astype(jnp.int32)
+        py = state.player_positions[i, 1].astype(jnp.int32)
         return jax.lax.dynamic_update_slice(
-            img, player_sprite, (py * tile_px, px * tile_px, 0)
+            img, player_sprite, (py * tile_px, px * tile_px, jnp.int32(0))
         )
 
     composited: jnp.ndarray = jax.lax.fori_loop(0, num_players, _stamp_player, image)
